Add JS unit tests
