+++ plugins/ui/src/js/src/ReactPanel.test.tsx
@@ -8,24 +8,6 @@ import {
 } from './ReactPanelManager';
 import { ReactPanelProps } from './layout/LayoutUtils';
 
-// Mock LayoutUtils, useListener, and PanelEvent from @deephaven/dashboard package
-const mockLayout = { root: {}, eventHub: {} };
-jest.mock('@deephaven/dashboard', () => {
-  const DashboardActual = jest.requireActual('@deephaven/dashboard');
-  return {
-    ...DashboardActual,
-    LayoutUtils: {
-      getComponentName: jest.fn(),
-      openComponent: jest.fn(),
-      closeComponent: jest.fn(),
-    },
-    useLayoutManager: jest.fn(() => mockLayout),
-    useListener: jest.fn(),
-    __esModule: true,
-    default: jest.fn(),
-  };
-});
-
 const mockPanelId = 'test-panel-id';
 jest.mock('shortid', () => jest.fn(() => mockPanelId));
