+++ plugins/ui/src/js/src/layout/Dashboard.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import type { StackElementProps } from './LayoutUtils';
+import type { DashboardElementProps } from './LayoutUtils';
 import { ParentItemContext, useParentItem } from './ParentItemContext';
 
-function Dashboard({ children }: StackElementProps): JSX.Element | null {
+function Dashboard({ children }: DashboardElementProps): JSX.Element | null {
   const parent = useParentItem();
 
   return (
